Preventor ported

+++ src/Views/ProcessView/ProcessInfoView/ProcessInfoView.vala
@@ -1,4 +1,5 @@
 public class Monitor.ProcessInfoView : Gtk.Box {
+    private Preventor preventor;
     private ProcessInfoIOStats process_info_io_stats = new ProcessInfoIOStats ();
 
     private Process _process;
@@ -21,6 +22,7 @@ public class Monitor.ProcessInfoView : Gtk.Box {
                 process_info_io_stats.hide ();
                 end_process_button.hide ();
                 kill_process_button.hide ();
+                preventor.hide ();
             } else {
                 _process.fd_permission_error.connect (show_permission_error_infobar);
 
@@ -37,6 +39,7 @@ public class Monitor.ProcessInfoView : Gtk.Box {
                 process_info_io_stats.set_visible (true);
                 end_process_button.show ();
                 kill_process_button.show ();
+                preventor.show();
             }
         }
     }
@@ -114,15 +117,24 @@ public class Monitor.ProcessInfoView : Gtk.Box {
         process_action_bar.append (kill_process_button);
 
         append (grid);
-        append (process_action_bar);
+
+        preventor = new Preventor (process_action_bar, "process_action_bar");
+        grid.attach (preventor, 0, 5, 1, 1);
+        preventor.hide ();
 
 
         kill_process_button.clicked.connect (() => {
-            process.kill (); // maybe add a toast that process killed
+            preventor.set_prevention (_("Confirm kill of the process?"));
+            preventor.confirmed.connect ((is_confirmed) => {
+                if (is_confirmed) process.kill (); // maybe add a toast that process killed
+            });
         });
 
         end_process_button.clicked.connect (() => {
-            process.end (); // maybe add a toast that process ended
+            preventor.set_prevention (_("Confirm end of the process?"));
+            preventor.confirmed.connect ((is_confirmed) => {
+                if (is_confirmed) process.end (); // maybe add a toast that process ended
+            });
         });
 
     }
